+++ packages/evian-tracking/src/sensor.rs
@@ -1,6 +1,8 @@
 use core::cell::RefCell;
 
 use alloc::{rc::Rc, vec::Vec};
+use evian_math::Angle;
+use vexide::devices::smart::imu::{InertialError, InertialSensor};
 use vexide::devices::{
     PortError,
     adi::AdiEncoder,
@@ -126,5 +128,39 @@ impl<T: RotarySensor> RotarySensor for Rc<RefCell<T>> {
 
     fn position(&self) -> Result<Position, Self::Error> {
         self.borrow().position()
+    }
+}
+
+/// A "gyroscope," or a sensor that measures the robot's heading and angular velocity
+pub trait Gyro {
+    /// The error returned when [`Gyro::heading`] or [`Gyro::angular_velocity`] fails. This
+    /// describes a hardware error.
+    type Error;
+
+    /// Returns the heading of the robot as an [`Angle`]
+    fn heading(&self) -> Result<Angle, Self::Error>;
+    /// Returns the horizontal angular velocity
+    fn angular_velocity(&self) -> Result<f64, Self::Error>;
+}
+
+impl Gyro for InertialSensor {
+    type Error = InertialError;
+
+    fn heading(&self) -> Result<Angle, Self::Error> {
+        InertialSensor::heading(self).map(Angle::from_radians)
+    }
+
+    fn angular_velocity(&self) -> Result<f64, Self::Error> {
+        InertialSensor::gyro_rate(self).map(|rate| rate.z)
+    }
+}
+impl<const N: usize> Gyro for [InertialSensor; N] {
+    type Error = [Option<InertialError>; N];
+
+    fn angular_velocity(&self) -> Result<f64, Self::Error> {
+        Ok(1.0)
+    }
+    fn heading(&self) -> Result<Angle, Self::Error> {
+        Ok(Angle::from_degrees(1.0))
     }
 }
